requirements.txt
Django==1.8.5
django-cors-headers==1.1.0
djangorestframework==3.3.3
Pillow==3.2.0
